Initial Implementation of Input System to be built upon
