+++ src/DeviceInput/index.ts
@@ -1,3 +1,1 @@
-export * from "./Profile/deviceSourceManager";
-export * from "./Systems/webDeviceInputSystem";
 export * from "./deviceInputSystem";
